data_engine/video-reasoning/ingest/video-embedder/requirements.txt
# Cloud events
cloudevents==1.10.1

# HTTP requests
requests==2.31.0

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0

# OpenTelemetry for tracing
opentelemetry-api
opentelemetry-sdk

# Vast Runtime (may be custom package)
# vast_runtime

